Combined ByteBufferPool
Move Bucket out of interface into abstract pool
